+++ lib/page/account/import/importAccountForm.dart
@@ -1,4 +1,4 @@
-import 'package:encointer_wallet/common/components/accountAdvanceOption.dart';
+import 'package:encointer_wallet/common/components/accountAdvanceOptionParams.dart';
 import 'package:encointer_wallet/common/components/encointerTextFormField.dart';
 import 'package:encointer_wallet/common/components/gradientElements.dart';
 import 'package:encointer_wallet/store/account/account.dart';
